+++ packages/mrm-task-lint-staged/index.js
@@ -1,10 +1,17 @@
 // @ts-check
-const { packageJson, install, getExtsFromCommand } = require('mrm-core');
+const {
+	packageJson,
+	install,
+	getExtsFromCommand,
+	uninstall,
+} = require('mrm-core');
+const { isUsingYarnBerry } = require('mrm-core/src/npm');
 const { castArray } = require('lodash');
+const husky = require('husky');
 
 const packages = {
 	'lint-staged': '>=10',
-	husky: '>=4',
+	husky: '>=6',
 };
 
 /**
@@ -172,19 +179,41 @@ module.exports = function task({ lintStagedRules }) {
 	pkg
 		// Remove husky 0.14 config
 		.unset('scripts.precommit')
+		// Remove husky 4 config
+		.unset('husky')
+		// Remove simple-git-hooks
+		.unset('simple-git-hooks')
 		// Add new config
 		.merge({
-			husky: {
-				hooks: {
-					'pre-commit': 'lint-staged',
-				},
-			},
 			'lint-staged': rules,
-		})
-		.save();
+		});
 
+	if (isUsingYarnBerry()) {
+		// Yarn 2 doesn't support `prepare` lifecycle yet
+		// https://yarnpkg.com/advanced/lifecycle-scripts
+		pkg.appendScript('postinstall', 'husky install');
+		if (!pkg.get('private')) {
+			// In case package isn't private, pinst ensures that postinstall
+			// is disabled on publish
+			pkg
+				.appendScript('prepublishOnly', 'pinst --disable')
+				.appendScript('postpublish', 'pinst --enable');
+			packages.pinst = '>=2';
+		}
+	} else {
+		// npm, Yarn 1, pnpm
+		pkg.appendScript('prepare', 'husky install');
+	}
+
+	pkg.save();
+
+	uninstall('simple-git-hooks');
 	// Install dependencies
 	install(packages);
+	// Install husky
+	husky.install();
+	// Set lint-staged config
+	husky.add('.husky/pre-commit', 'npx lint-staged');
 };
 
 module.exports.description = 'Adds lint-staged';