+++ packages/mrm-task-lint-staged/index.spec.js
@@ -1,10 +1,18 @@
 jest.mock('fs');
 jest.mock('mrm-core/src/npm', () => ({
+	uninstall: jest.fn(),
 	install: jest.fn(),
+	isUsingYarnBerry: jest.fn(),
+}));
+jest.mock('husky', () => ({
+	install: jest.fn(),
+	add: jest.fn(),
 }));
 
-const { install } = require('mrm-core');
+const { install, uninstall } = require('mrm-core');
+const { isUsingYarnBerry } = require('mrm-core/src/npm');
 const { getTaskOptions } = require('mrm');
+const husky = require('husky');
 const vol = require('memfs').vol;
 const task = require('./index');
 
@@ -26,6 +34,10 @@ beforeEach(() => {
 afterEach(() => {
 	vol.reset();
 	install.mockClear();
+	uninstall.mockClear();
+	isUsingYarnBerry.mockClear();
+	husky.install.mockClear();
+	husky.add.mockClear();
 	console.log = console$log;
 });
 
@@ -56,7 +68,13 @@ it('should add Prettier if project depends on it', async () => {
 	task(await getTaskOptions(task));
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(uninstall).toBeCalledWith('simple-git-hooks');
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
+	expect(husky.install).toHaveBeenCalled();
+	expect(husky.add).toBeCalledWith('.husky/pre-commit', 'npx lint-staged');
 });
 
 it('should add Prettier and ESLint', async () => {
@@ -128,7 +146,11 @@ it('should add ESLint if project depends on it', async () => {
 	task(await getTaskOptions(task));
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(uninstall).toBeCalledWith('simple-git-hooks');
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
 });
 
 it('should use default JS extension if eslint command has no --ext key', async () => {
@@ -238,7 +260,10 @@ it('should add stylelint if project depends on it', async () => {
 	task(await getTaskOptions(task));
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
 });
 
 it('should use a custom stylelint extension', async () => {
@@ -277,7 +302,11 @@ it('should add a custom rule', async () => {
 	);
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(uninstall).toBeCalledWith('simple-git-hooks');
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
 });
 
 it('should update an existing rule', async () => {
@@ -297,7 +326,11 @@ it('should update an existing rule', async () => {
 	task(await getTaskOptions(task, false));
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(uninstall).toBeCalledWith('simple-git-hooks');
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
 });
 
 it('should merge rules with the same pattern', async () => {
@@ -325,7 +358,11 @@ it('should merge rules with the same pattern', async () => {
 	);
 
 	expect(vol.toJSON()).toMatchSnapshot();
-	expect(install).toBeCalledWith({ 'lint-staged': '>=10', husky: '>=4' });
+	expect(uninstall).toBeCalledWith('simple-git-hooks');
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
 });
 
 it('should remove husky 0.14 config from package.json', async () => {
@@ -343,5 +380,67 @@ it('should remove husky 0.14 config from package.json', async () => {
 
 	task(await getTaskOptions(task));
 
+	expect(vol.toJSON()).toMatchSnapshot();
+});
+
+it('should remove husky 4 config from package.json', async () => {
+	vol.fromJSON({
+		'/package.json': stringify({
+			name: 'unicorn',
+			husky: {
+				hooks: {
+					'pre-commit': 'lint-staged',
+				},
+			},
+			devDependencies: {
+				eslint: '*',
+			},
+		}),
+	});
+
+	task(await getTaskOptions(task));
+
+	expect(vol.toJSON()).toMatchSnapshot();
+});
+
+it('should use Yarn 2 specific lifecycle', async () => {
+	vol.fromJSON({
+		'/package.json': stringify({
+			private: true,
+			name: 'unicorn',
+			devDependencies: {
+				eslint: '*',
+			},
+		}),
+	});
+
+	isUsingYarnBerry.mockReturnValue(true);
+	task(await getTaskOptions(task));
+
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+	});
+	expect(vol.toJSON()).toMatchSnapshot();
+});
+
+it('should use Yarn 2 specific lifecycle (public)', async () => {
+	vol.fromJSON({
+		'/package.json': stringify({
+			name: 'unicorn',
+			devDependencies: {
+				eslint: '*',
+			},
+		}),
+	});
+
+	isUsingYarnBerry.mockReturnValue(true);
+	task(await getTaskOptions(task));
+
+	expect(install).toBeCalledWith({
+		'lint-staged': '>=10',
+		husky: '>=6',
+		pinst: '>=2',
+	});
 	expect(vol.toJSON()).toMatchSnapshot();
 });