+++ packages/mrm-task-dependabot/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/sapegin/mrm/compare/mrm-task-dependabot@1.3.0...mrm-task-dependabot@1.3.1) (2021-04-01)
+
+**Note:** Version bump only for package mrm-task-dependabot
+
+
+
+
+
+# [1.3.0](https://github.com/sapegin/mrm/compare/mrm-task-dependabot@1.2.6...mrm-task-dependabot@1.3.0) (2021-02-25)
+
+
+### Features
+
+* Add more fields to package.json ([#134](https://github.com/sapegin/mrm/issues/134)) ([d80840a](https://github.com/sapegin/mrm/commit/d80840a5e771976ef38cdf8a3b535a412e1097f6))
+
+
+
+
+
+## [1.2.6](https://github.com/sapegin/mrm/compare/mrm-task-dependabot@1.2.5...mrm-task-dependabot@1.2.6) (2021-02-09)
+
+**Note:** Version bump only for package mrm-task-dependabot
+
+
+
+
+
 ## [1.2.5](https://github.com/sapegin/mrm/compare/mrm-task-dependabot@1.2.4...mrm-task-dependabot@1.2.5) (2021-02-08)
 
 **Note:** Version bump only for package mrm-task-dependabot