+++ packages/mrm-task-lint-staged/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@4.0.0...mrm-task-lint-staged@5.0.0) (2021-04-01)
+
+
+### Features
+
+* **lint-staged:** Use husky 6 ([#152](https://github.com/sapegin/mrm/issues/152)) ([133fe08](https://github.com/sapegin/mrm/commit/133fe08b0895b0c994c55d39e0f43af0672fe1f9))
+
+
+### BREAKING CHANGES
+
+* **lint-staged:** The lint-staged task will migrate simple-git-hooks back to husky 6, the existing simple-git-hooks dependency will be removed.
+
+
+
+
+
+# [4.0.0](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@3.1.0...mrm-task-lint-staged@4.0.0) (2021-03-22)
+
+
+### Features
+
+* Replace husky with simple-git-hooks ([#149](https://github.com/sapegin/mrm/issues/149)) ([8ecd2ce](https://github.com/sapegin/mrm/commit/8ecd2ce0816fb06c395276250fa85dea6f93686d))
+
+
+### BREAKING CHANGES
+
+* The lint-staged task will migrate `husky` to `simple-git-hooks`, the existing `husky` dependency will be removed.
+
+
+
+
+
+# [3.1.0](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@3.0.15...mrm-task-lint-staged@3.1.0) (2021-02-25)
+
+
+### Features
+
+* Add more fields to package.json ([#134](https://github.com/sapegin/mrm/issues/134)) ([d80840a](https://github.com/sapegin/mrm/commit/d80840a5e771976ef38cdf8a3b535a412e1097f6))
+
+
+
+
+
+## [3.0.15](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@3.0.14...mrm-task-lint-staged@3.0.15) (2021-02-20)
+
+
+### Bug Fixes
+
+* Fix broken lint-staged action by explicitly pining Husky to 4.x ([#146](https://github.com/sapegin/mrm/issues/146)) ([29d2194](https://github.com/sapegin/mrm/commit/29d2194d46c4160d942f559c8511cfbb8b52704e)), closes [#145](https://github.com/sapegin/mrm/issues/145)
+
+
+
+
+
+## [3.0.14](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@3.0.13...mrm-task-lint-staged@3.0.14) (2021-02-09)
+
+**Note:** Version bump only for package mrm-task-lint-staged
+
+
+
+
+
 ## [3.0.13](https://github.com/sapegin/mrm/compare/mrm-task-lint-staged@3.0.12...mrm-task-lint-staged@3.0.13) (2021-02-08)
 
 **Note:** Version bump only for package mrm-task-lint-staged