+++ packages/mrm/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.2](https://github.com/sapegin/mrm/compare/mrm@2.6.1...mrm@2.6.2) (2021-04-01)
+
+**Note:** Version bump only for package mrm
+
+
+
+
+
+## [2.6.1](https://github.com/sapegin/mrm/compare/mrm@2.6.0...mrm@2.6.1) (2021-03-22)
+
+**Note:** Version bump only for package mrm
+
+
+
+
+
+# [2.6.0](https://github.com/sapegin/mrm/compare/mrm@2.5.19...mrm@2.6.0) (2021-02-25)
+
+
+### Features
+
+* Add more fields to package.json ([#134](https://github.com/sapegin/mrm/issues/134)) ([d80840a](https://github.com/sapegin/mrm/commit/d80840a5e771976ef38cdf8a3b535a412e1097f6))
+
+
+
+
+
+## [2.5.19](https://github.com/sapegin/mrm/compare/mrm@2.5.18...mrm@2.5.19) (2021-02-20)
+
+**Note:** Version bump only for package mrm
+
+
+
+
+
+## [2.5.18](https://github.com/sapegin/mrm/compare/mrm@2.5.17...mrm@2.5.18) (2021-02-09)
+
+**Note:** Version bump only for package mrm
+
+
+
+
+
+## [2.5.17](https://github.com/sapegin/mrm/compare/mrm@2.5.16...mrm@2.5.17) (2021-02-09)
+
+**Note:** Version bump only for package mrm
+
+
+
+
+
 ## [2.5.16](https://github.com/sapegin/mrm/compare/mrm@2.5.15...mrm@2.5.16) (2021-02-08)
 
 **Note:** Version bump only for package mrm